+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/pyaud/compare/v5.0.1...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Add `pyaud.plugins.Subprocesses`
 - Add `pyaud.messages`
 
 ### Changed